chore: files changed crates/tinyflows-catalog/src/import/n8n/node_mapping.rs,crates/tinyflows-catalo
Auto-committed-on: dragonfly
